refactor: diary list 조회 test code 분리
